Server/requirements.txt
django
pyjwt
gunicorn
requests
Pillow
filetype
djangorestframework
markdown
django-filter
drf_yasg

# Tests
mock